update- moved some more functions over to a file & corrected some css errors
