+++ src/CSS/page1.css
@@ -156,8 +156,8 @@ body
     margin-bottom: 10px;
     text-align: center;
     font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-    width: 80%;
-    height: 7%;
+    width: 60%;
+    height: 6%;
     top: 0;
     left: 0;
     transition-duration: 0.5s;
@@ -430,6 +430,7 @@ input.search-area
     border: none;
     padding: 5px;
     background-color: #8181817e;
+    color: black;
     border-radius: 5px;
     transition-duration: 0.4s;
     pointer-events: all;
@@ -437,6 +438,7 @@ input.search-area
 .p-d-category:hover, .p-d-vendor:hover, .p-d-type:hover
 {
     background-color: #b8b8b87e;
+    color: black;
 }
 
 .p-d-category, .p-d-vendor, .p-d-type, a:hover 
@@ -552,6 +554,12 @@ input.search-area
     transform: translate(76%);
 }
 
+.pan-main a
+{
+    color: black;
+    transition-duration: 0.4s;
+    text-decoration: none;
+}
 a
 {
     color: black;
