+++ src/JS/Functions/Order_Functions.js
@@ -77,7 +77,6 @@ export function DetailedView()
                             Address3 ={el.city} Address4={el.suburb} Address5={el.postal_code}
                             />))
                         }
-                        
                     }, 10);    
                 }
                 else
