+++ src/components/Page2.js
@@ -3,11 +3,12 @@ import { flushSync } from 'react-dom';
 import { createRoot } from 'react-dom/client';
 import Chart from "chart.js/auto";
 import Queue_details from './semi-components/queue-details';
-import Detailed_queue from './semi-components/Queue/detailed_queue';
 import $ from 'jquery';
 
 import '../CSS/page1.css';
 
+import { Filter_Pagintation } from '../JS/Functions/Queue_Functions';
+import { DetailedView } from '../JS/Functions/Queue_Functions';
 
 function Page2(props)
 {
@@ -72,7 +73,6 @@ function Page2(props)
             {
                 for(let i = 0 ; i < queue_items.length; i++)
                 {
-                    console.log(i);
                     queue_items[i].style.cursor = "pointer";
                     queue_items[i].style.pointerEvents = "";
                     queue_items[i].addEventListener("click", () => 
@@ -196,345 +196,6 @@ function Page2(props)
             .fail( function(xhr) { alert(xhr.responseText); });
             Filter_Pagintation(1);
         });  
-    
-        function DetailedView()
-        {
-            let queue = document.querySelector(".queue");
-            let pan = document.querySelectorAll(".pan");
-            for(let i = 0; i < pan.length; i++)
-            {
-                pan[i].addEventListener("click", () =>
-                {
-                    let id = pan[i].querySelector(".p-d-id").innerHTML;
-                    /*  API  */
-                    const api_key = localStorage.getItem('api_key');
-                    $.ajaxSetup({ headers: { 'Authorization': 'ApiKey ' + api_key} });
-                    $.get("http://localhost:8080/api/queue/" + id, [], [], 'json')
-                    .done(function(_data) 
-                    {   
-                        console.log(_data);
-                        if(document.querySelector(".details") != null)
-                        //div already exists, remove it, and create another
-                        {
-                            document.querySelector(".details").remove();
-                            let details = document.createElement('div');
-                            details.className = "details";
-                            queue.appendChild(details);
-
-                            let rot = createRoot(details);
-                            rot.render( <Detailed_queue key={`${_data.title}_${i}`} Queue_Status={_data.status} Queue_Description={_data.description}
-                            Queue_Type={_data.queue_type} Queue_Instruction={_data.instruction} Queue_ID={_data.id} Created_At={_data.created_at} 
-                            Updated_At={_data.updated_at}
-                            />)
-                            
-                        }
-                        else 
-                        //create new div
-                        {
-                            let details = document.createElement('details');
-                            queue.appendChild(details);
-                            let rot = createRoot(details);
-                            rot.render( <Detailed_queue key={`${_data.title}_${i}`} Queue_Status={_data.status} Queue_Description={_data.description}
-                            Queue_Type={_data.queue_type} Queue_Instruction={_data.instruction} Queue_ID={_data.id} Created_At={_data.created_at} 
-                            Updated_At={_data.updated_at}
-                            />)
-                        }
-                    })
-                    .fail( function(xhr) 
-                    {
-                        alert(xhr.responseText);
-                    });
-                    setTimeout(() =>
-                    {
-                        let filter = document.querySelector(".filter");
-                        let main = document.querySelector(".main");
-                        let navbar = document.getElementById("navbar");
-                        let details = document.querySelector(".details");
-
-                        filter.style.animation = "Fadeout 0.5s ease-out";
-                        main.style.animation = "Fadeout 0.5s ease-out";
-                        navbar.style.animation = "Fadeout 0.5s ease-out";
-                        filter.style.display = "none";
-                        main.style.display = "none";
-                        navbar.style.display = "none";
-                        details.style.display = "block";
-                    }, 50);
-                });
-            } 
-        }
-
-        /* Script to automatically format the number of elements on each page */
-        const content = document.querySelector('.center'); 
-        const paginationContainer = document.createElement('div');
-        const paginationDiv = document.body.appendChild(paginationContainer);
-        paginationContainer.classList.add('pagination');
-        content.appendChild(paginationContainer);
-
-        function Filter_Pagintation(index)
-        {
-            if(index == 1)
-            {
-
-                let next = document.getElementById("next");
-                let ahead = index + 1;
-                for(let i = 0; i < filter_img.length; i++)
-                {
-                    if(filter_img[i].style.display == "block")
-                    {
-                        let type = filter_img[i].nextSibling.className;
-                        const api_key = localStorage.getItem('api_key');
-                        $.ajaxSetup({ headers: { 'Authorization': 'ApiKey ' + api_key} });
-                        $.get("http://localhost:8080/api/queue/filter?type=" + type + "?page="+ ahead, [], [], 'json')
-                        .done(function( _data) 
-                        {
-                            console.log(_data);
-                            if(_data.length < 10)
-                            {
-                                next.disabled = true;
-                                next.style.cursor = "not-allowed";
-                            }
-                            if(_data == "") { let next = document.getElementById("next"); next.style.cursor = "not-allowed"; next.disabled = true; } 
-                            else 
-                            {
-                                if(document.querySelector(".pan-main") != null)
-                                {
-                                    document.querySelector(".pan-main").remove();
-                                    let pan_main = document.createElement('div');
-                                    let main_elements = document.querySelector(".main-elements");
-                                    pan_main.className = "pan-main";
-                                    main_elements.appendChild(pan_main);
-
-                                    let root = createRoot(pan_main);
-                                    flushSync(() => 
-                                    {
-                                        root.render(_data.map((el, i) => <Queue_details key={`${el.title}_${i}`} 
-                                        />))
-                                    });
-                                    setTimeout(() => { DetailedView();}, 200);
-                                }
-                                else 
-                                {
-                                    let pan_main = document.createElement('div');
-                                    let main_elements = document.querySelector(".main-elements");
-                                    pan_main.className = "pan-main";
-                                    main_elements.appendChild(pan_main);
-
-                                    let root = createRoot(pan_main);
-                                    flushSync(() => 
-                                    {
-                                        root.render(_data.map((el, i) => <Queue_details key={`${el.title}_${i}`} 
-                                        Queue_Type={el.queue_type} Queue_Instruction={el.instruction} Queue_Status={el.status} Queue_ID={el.id}
-                                        />))
-                                    });
-                                    setTimeout(() => { DetailedView();}, 200);
-                                }
-                            }
-                        })
-                        .fail( function(xhr) { alert(xhr.responseText); });
-                    }
-                }
-                
-            }
-
-            /* Check done to remove old elements if they exist */
-            if(document.getElementById("next") != null && document.getElementById("prev") != null && document.getElementById("hod") != null)
-            //If they exist remove them, and create new based on the new index value
-            {
-                document.getElementById("next").remove();
-                document.getElementById("prev").remove();
-                document.getElementById("hod").remove();
-
-                const pageButton = document.createElement('button');
-                pageButton.id = "hod";
-                pageButton.className = "active";
-                pageButton.innerHTML = index;
-                paginationDiv.appendChild(pageButton);
-
-                const nextPage = document.createElement('button');
-                nextPage.id = "next";
-                nextPage.innerHTML = "→";
-                paginationDiv.appendChild(nextPage);
-
-                const prevPage = document.createElement('button');
-                prevPage.id = "prev";
-                prevPage.innerHTML = "←";
-                paginationDiv.appendChild(prevPage);
-                if(index == 1) { prevPage.disabled = true; prevPage.style.cursor = "not-allowed"; }
-                else if(index > 1) { prevPage.style.cursor = "pointer"; prevPage.disabled = false; nextPage.disabled = false; }
-                else if(index <= 1) {prevPage.disabled = true; prevPage.style.cursor = "not-allowed"; }
-
-                nextPage.addEventListener("click", () =>
-                {
-                    let category = document.querySelector(".category").innerHTML;
-                    let type = document.querySelector(".type").innerHTML;
-                    let vendor = document.querySelector(".vendor").innerHTML;
-
-                    index = index + 1;
-                    /*  API  */
-                    const api_key = localStorage.getItem('api_key');
-                    $.ajaxSetup({ headers: { 'Authorization': 'ApiKey ' + api_key} });
-                    $.get("http://localhost:8080/api/products/filter?type=" + type + "&vendor=" + vendor + "&category=" + category + "&page=" + index, [], [])
-                    .done(function( _data) 
-                    {
-                        console.log(_data);
-
-                        document.querySelector(".pan-main").remove();
-                        let div = document.createElement("div");
-                        div.className = "pan-main";
-                        div.id = "pan-main";
-                        let main = document.querySelector(".main-elements");
-                        main.appendChild(div);
-                        let root = createRoot(div);
-                        flushSync(() => 
-                        { 
-                            root.render(_data.map((el, i) => <Queue_details key={`${el.title}_${i}`}
-                            Queue_Type={el.queue_type} Queue_Instruction={el.instruction} Queue_Status={el.status} Queue_ID={el.id}
-                            />))
-                        });
-                        
-                    })
-                    .fail( function(xhr) { alert(xhr.responseText); });
-
-                    let ahead = index + 1;
-                    /*  API  */
-                    $.get("http://localhost:8080/api/products/filter?type=" + type + "&vendor=" + vendor + "&category=" + category + "&page=" + ahead, [], [])
-                    .done(function( _data) 
-                    {
-                        console.log(_data);
-                        if(_data == "") { let next = document.getElementById("next"); next.style.cursor = "not-allowed"; next.disabled = true; }  
-                    })
-                    .fail( function(xhr) { alert(xhr.responseText); });
-
-                    Filter_Pagintation(index);
-                    setTimeout(() => { DetailedView();}, 200);
-                });
-
-                prevPage.addEventListener("click", () =>
-                {
-                    let category = document.querySelector(".category").innerHTML;
-                    let type = document.querySelector(".type").innerHTML;
-                    let vendor = document.querySelector(".vendor").innerHTML;
-                    index = index - 1;
-                    const api_key = localStorage.getItem('api_key');
-                    $.ajaxSetup({ headers: { 'Authorization': 'ApiKey ' + api_key} });
-                    $.get("http://localhost:8080/api/products/filter?type=" + type + "&vendor=" + vendor + "&category=" + category + "&page=" + index, [], [])
-                    .done(function( _data) 
-                    {
-                        console.log(_data);
-                        document.querySelector(".pan-main").remove();
-                        let div = document.createElement("div");
-                        div.className = "pan-main";
-                        div.id = "pan-main";
-                        let main = document.querySelector(".main-elements");
-                        main.appendChild(div);
-                        let root = createRoot(div);
-                        flushSync(() => 
-                        { 
-                            root.render(_data.map((el, i) => <Queue_details key={`${el.title}_${i}`}
-                            Queue_Type={el.queue_type} Queue_Instruction={el.instruction} Queue_Status={el.status} Queue_ID={el.id}
-                            />))
-                        });
-
-                    
-                    })
-                    .fail( function(xhr) { alert(xhr.responseText); });
-                    Filter_Pagintation(index);
-                    setTimeout(() => { DetailedView();}, 200);
-                });
-            }
-            else 
-            //If they dont exist create new ones 
-            {
-                const pageButton = document.createElement('button');
-                pageButton.id = "hod";
-                pageButton.className = "active";
-                pageButton.innerHTML = index;
-                paginationDiv.appendChild(pageButton);
-
-                const nextPage = document.createElement('button');
-                nextPage.id = "next";
-                nextPage.innerHTML = "→";
-                paginationDiv.appendChild(nextPage);
-
-                const prevPage = document.createElement('button');
-                prevPage.id = "prev";
-                prevPage.innerHTML = "←";
-                paginationDiv.appendChild(prevPage);
-
-                if(index == 1) { prevPage.disabled = true; prevPage.style.cursor = "not-allowed"; }
-                else if(index > 1) { prevPage.style.cursor = "pointer"; prevPage.disabled = false; nextPage.disabled = false; }
-                else if(index <= 1) {prevPage.disabled = true; prevPage.style.cursor = "not-allowed"; }
-
-                nextPage.addEventListener("click", () =>
-                {
-                    let category = document.querySelector(".category").innerHTML;
-                    let type = document.querySelector(".type").innerHTML;
-                    let vendor = document.querySelector(".vendor").innerHTML;
-
-                    index = index + 1;
-
-                    const api_key = localStorage.getItem('api_key');
-                    $.ajaxSetup({ headers: { 'Authorization': 'ApiKey ' + api_key} });
-                    $.get("http://localhost:8080/api/products/filter?type=" + type + "&vendor=" + vendor + "&category=" + category + "&page=" + index, [], [])
-                    .done(function( _data) 
-                    {
-                        console.log(_data);
-
-                        document.querySelector(".pan-main").remove();
-                        let div = document.createElement("div");
-                        div.className = "pan-main";
-                        div.id = "pan-main";
-                        let main = document.querySelector(".main-elements");
-                        main.appendChild(div);
-                        let root = createRoot(div);
-                        flushSync(() => 
-                        { 
-                            root.render(_data.map((el, i) => <Queue_details key={`${el.title}_${i}`} Queue_Updated_At={el.updated_at} Queue_Creation_Date={el.created_at} 
-                            Queue_Type={el.queue_type} Queue_Instruction={el.instruction} Queue_Status={el.status} Queue_ID={el.id}
-                            />)) 
-                        });
-
-                    })
-                    .fail( function(xhr) { alert(xhr.responseText); });
-                    Filter_Pagintation(index);
-                    setTimeout(() => { DetailedView();}, 200);
-                });
-
-                prevPage.addEventListener("click", () =>
-                {
-                    let category = document.querySelector(".category").innerHTML;
-                    let type = document.querySelector(".type").innerHTML;
-                    let vendor = document.querySelector(".vendor").innerHTML;
-                    index = index - 1;
-
-                    const api_key = localStorage.getItem('api_key');
-                    $.ajaxSetup({ headers: { 'Authorization': 'ApiKey ' + api_key} });
-                    $.get("http://localhost:8080/api/products/filter?type=" + type + "&vendor=" + vendor + "&category=" + category + "&page=" + index, [], [])
-                    .done(function( _data) 
-                    {
-                        console.log(_data);
-
-                        document.querySelector(".pan-main").remove();
-                        let div = document.createElement("div");
-                        div.className = "pan-main";
-                        div.id = "pan-main";
-                        let main = document.querySelector(".main-elements");
-                        main.appendChild(div);
-                        let root = createRoot(div);
-                        flushSync(() => 
-                        { 
-                            root.render(_data.map((el, i) => <Queue_details key={`${el.title}_${i}`}
-                            Queue_Type={el.queue_type} Queue_Instruction={el.instruction} Queue_Status={el.status} Queue_ID={el.id}
-                            />))
-                        });
-                        
-                    })
-                    .fail( function(xhr) { alert(xhr.responseText); });
-                    Filter_Pagintation(index);
-                    setTimeout(() => { DetailedView();}, 200);
-                });
-            } 
-        }
 
         /* Queue Information Requests */
         let graph_data = {};
